i think last commit before integration

+++ src/main/java/tn/esprit/ecocycletech/Service/AppareilsManagement/AppareilServiceImpl.java
@@ -3,6 +3,7 @@ package tn.esprit.ecocycletech.Service.AppareilsManagement;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tn.esprit.ecocycletech.Entity.AppareilsManagement.Appareil;
+import tn.esprit.ecocycletech.Entity.AppareilsManagement.Avis;
 import tn.esprit.ecocycletech.Repository.AppareilsManagement.IAppareilRepository;
 
 import java.util.List;
@@ -24,17 +25,19 @@ public class AppareilServiceImpl implements IAppareilService {
 
     @Override
     public boolean deleteAppareil(int id) {
-        appareilRepository.deleteById(id);
+        if (appareilRepository.existsById(id)) {
+            appareilRepository.deleteById(id);
+            return true;
+        }
         return false;
     }
 
-    // Récupérer tous les appareils
+
     @Override
     public List<Appareil> getAllAppareils() {
         return appareilRepository.findAll();
     }
 
-    // Mettre à jour un appareil existant
     @Override
     public Appareil updateAppareil(int id, Appareil appareil) {
         Optional<Appareil> existingAppareilOpt = appareilRepository.findById(id);
@@ -44,7 +47,6 @@ public class AppareilServiceImpl implements IAppareilService {
             existingAppareil.setCategorie(appareil.getCategorie());
             existingAppareil.setEtatAppareil(appareil.getEtatAppareil());
             existingAppareil.setMarque(appareil.getMarque());
-            existingAppareil.setQuantite(appareil.getQuantite());
             existingAppareil.setPrix(appareil.getPrix());
             existingAppareil.setDescription(appareil.getDescription());
             existingAppareil.setImageurl(appareil.getImageurl());
@@ -52,4 +54,15 @@ public class AppareilServiceImpl implements IAppareilService {
             return appareilRepository.save(existingAppareil);
         }
         return null;
-}}
+}
+    public double calculerNoteMoyenne(Appareil appareil) {
+        if (appareil.getAvis() == null || appareil.getAvis().isEmpty()) {
+            return 0.0;
+        }
+        return appareil.getAvis().stream()
+                .mapToInt(Avis::getRating)
+                .average()
+                .orElse(0.0);
+    }
+
+}
